# features/done/web-ui/openapi-v0.1.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "fallen-8-core-apiApp | v0.1",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://127.0.0.1:5078/"
    }
  ],
  "paths": {
    "/status": {
      "get": {
        "tags": [
          "Admin"
        ],
        "summary": "Gets the current status of the Fallen-8 database",
        "parameters": [
          {
            "name": "api-version",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "X-Version",
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Returns the database status information",
            "content": {
              "application/json; ver=0.1": {
                "schema": {
                  "$ref": "#/components/schemas/StatusREST"
                }
              },
              "application/json": { }
            }
          }
        }
      }
    },
...
    "/load": {
      "put": {
        "tags": [
          "Admin"
        ],
        "summary": "Loads a Fallen-8 database from a saved file",
        "description": "Sample request:\r\n\r\n    PUT /load\r\n    {\r\n       \"startServices\": true,\r\n       \"saveGameLocation\": \"C:/Fallen8/database.f8s\"\r\n    }",
        "parameters": [
          {
            "name": "api-version",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "X-Version",
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "description": "Load specification including file path and service start options",
          "content": {
            "application/json; ver=0.1": {
              "schema": {
                "$ref": "#/components/schemas/LoadSpecification"
              }
            },
            "application/*+json; ver=0.1": {
              "schema": {
                "$ref": "#/components/schemas/LoadSpecification"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Database loaded successfully"
          },
          "400": {
            "description": "Invalid load specification or file not found",
            "content": {
              "text/plain; ver=0.1": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json; ver=0.1": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json; ver=0.1": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "500": {
            "description": "The load transaction was rolled back and did not complete"
          }
        }
      }
    },
    "/save": {
      "put": {
        "tags": [
          "Admin"
        ],
        "summary": "Saves the current database state to a file",
        "description": "Sample request:\r\n\r\n    PUT /save\r\n    {\r\n       \"saveGameLocation\": \"C:/Fallen8/database.f8s\",\r\n       \"savePartitions\": 8\r\n    }\r\n\r\nBoth parameters are optional. If not provided, defaults to the base directory with filename \"Temp.f8s\" and optimal partition count.\r\nThe save is recorded in the save-game registry (feature save-games); the response is the\r\ncreated entry, whose \"location\" field is the path the database was saved to.",
        "parameters": [
          {
            "name": "api-version",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "X-Version",
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "description": "Save specification including file path and partition options (both optional)",
          "content": {
            "application/json; ver=0.1": {
              "schema": {
                "$ref": "#/components/schemas/SaveSpecification"
              }
            },
            "application/*+json; ver=0.1": {
              "schema": {
                "$ref": "#/components/schemas/SaveSpecification"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Returns the created save-game registry entry",
            "content": {
              "application/json; ver=0.1": {
                "schema": {
                  "$ref": "#/components/schemas/SaveGameREST"
                }
              },
              "application/json": { }
            }
          },
          "400": {
            "description": "Invalid save specification",
            "content": {
              "application/json; ver=0.1": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": { }
            }
          },
          "500": {
            "description": "The save transaction was rolled back and did not complete",
            "content": {
              "application/json": { }
            }
          }
        }
      }
    },
...
    "/vertex/count": {
      "get": {
        "tags": [
          "Admin"
        ],
        "summary": "Gets the total number of vertices in the database",
        "parameters": [
          {
            "name": "api-version",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "X-Version",
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Returns the number of vertices",
            "content": {
              "application/json; ver=0.1": {
                "schema": {
                  "pattern": "^-?(?:0|[1-9]\\d*)$",
                  "type": [
                    "integer",
                    "string"
                  ],
                  "format": "int32"
                }
              },
              "application/json": { }
            }
          }
        }
      }
    },
    "/edge/count": {
      "get": {
        "tags": [
          "Admin"
        ],
        "summary": "Gets the total number of edges in the database",
        "parameters": [
          {
            "name": "api-version",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "X-Version",
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Returns the number of edges",
            "content": {
              "application/json; ver=0.1": {
                "schema": {
                  "pattern": "^-?(?:0|[1-9]\\d*)$",
                  "type": [
                    "integer",
                    "string"
                  ],
                  "format": "int32"
                }
              },
              "application/json": { }
            }
          }
        }
      }
    },
    "/service": {
      "post": {
        "tags": [
          "Admin"
        ],
        "summary": "Creates a new service based on the specified plugin",
        "parameters": [
          {
            "name": "api-version",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "X-Version",
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "description": "Plugin specification including type, ID and options",
          "content": {
            "application/json; ver=0.1": {
              "schema": {
                "$ref": "#/components/schemas/PluginSpecification"
              }
            },
            "application/*+json; ver=0.1": {
              "schema": {
                "$ref": "#/components/schemas/PluginSpecification"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Returns whether the service was successfully created",
            "content": {
              "application/json; ver=0.1": {
                "schema": {
                  "type": "boolean"
                }
              },
              "application/json": { }
            }
          },
          "400": {
            "description": "Invalid plugin specification",
            "content": {
              "application/json; ver=0.1": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": { }
            }
          }
        }
      }
    },
    "/service/{key}": {
      "delete": {
        "tags": [
          "Admin"
        ],
        "summary": "Deletes a service with the specified key",
        "parameters": [
          {
            "name": "key",
            "in": "path",
            "description": "The unique identifier of the service to delete",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "api-version",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "X-Version",
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Returns whether the service was successfully deleted",
            "content": {
              "application/json; ver=0.1": {
                "schema": {
                  "type": "boolean"
                }
              },
              "application/json": { }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "StatusREST": {
        "type": "object",
        "properties": {
          "usedMemory": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "description": "The amount of memory currently used by the database process in bytes",
            "format": "int64",
            "default": 1073741824,
            "example": 1073741824
          },
          "vertexCount": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "description": "The total number of vertices in the database",
            "format": "int32",
            "default": 10000,
            "example": 10000
          },
          "edgeCount": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "description": "The total number of edges in the database",
            "format": "int32",
            "default": 25000,
            "example": 25000
          },
          "indices": {
            "type": [
              "null",
              "array"
            ],
            "items": {
              "$ref": "#/components/schemas/IndexDescriptionREST"
            },
            "description": "The indices currently registered on this instance (id + plugin type) — the live\r\ninventory, available without running the budgeted statistics pass"
          },
          "availableIndexPlugins": {
            "type": [
              "null",
              "array"
            ],
            "items": {
              "type": "string"
            },
            "description": "List of available index plugins that can be used with the database",
            "example": [
              "DictionaryIndex",
              "SpatialIndex",
              "FullTextIndex"
            ]
          },
          "availablePathPlugins": {
            "type": [
              "null",
              "array"
            ],
            "items": {
              "type": "string"
            },
            "description": "List of available path-finding algorithm plugins",
            "example": [
              "Dijkstra",
              "AStar",
              "BellmanFord"
            ]
          },
          "availableAnalyticsPlugins": {
            "type": [
              "null",
              "array"
            ],
            "items": {
              "type": "string"
            },
            "description": "List of available graph-analytics algorithm plugins",
            "example": [
              "PAGERANK",
              "WCC",
              "LABELPROPAGATION",
              "DEGREE",
              "TRIANGLECOUNT"
            ]
          },
          "availableServicePlugins": {
            "type": [
              "null",
              "array"
            ],
            "items": {
              "type": "string"
            },
            "description": "List of available service plugins that can be started with the database",
            "example": [
              "ImportService",
              "ExportService",
              "AnalyticsService"
            ]
          },
          "apiKeyRequired": {
            "type": "boolean",
            "description": "True when this server has an API key configured, i.e. every endpoint outside the\r\nanonymous allowlist answers 401 without a valid credential. /status itself stays\r\nanonymous, so it doubles as the connection probe: a caller is authorized iff\r\n`!ApiKeyRequired || Authenticated`.",
            "example": false
          },
          "authenticated": {
            "type": "boolean",
            "description": "True when the request that produced this status carried a valid credential\r\n(see bool StatusREST.ApiKeyRequired for how clients combine the two).",
            "example": false
          }
        },
        "description": "Represents the current status of the Fallen-8 database",
        "example": {
          "usedMemory": 1073741824,
          "vertexCount": 10000,
          "edgeCount": 25000,
          "indices": [
            {
              "indexId": "nameIndex",
              "pluginType": "DictionaryIndex"
            }
          ],
          "availableIndexPlugins": [
            "DictionaryIndex",
            "SpatialIndex"
          ],
          "availablePathPlugins": [
            "Dijkstra",
            "AStar"
          ],
          "availableAnalyticsPlugins": [
            "PAGERANK",
            "WCC"
          ],
          "availableServicePlugins": [
            "ImportService",
            "ExportService"
          ],
          "apiKeyRequired": false,
          "authenticated": false
        }
      },
...
    }
  },
...
}